Made processes Windows compatible
`docker-compose run -d` is required on Windows because interactive mode is not yet supported there. I think it is actually safe to always run these commands detached since it is an automated tool and should not require any kind of interaction. I also added escaping to all arguments that might contain weird characters to make the commandline construction more resilient. Last but not least I DRYed the code a bit. That being said, I question the existence of the this class but this requires a bit more undertanding of the whole from my side to truly judge.