Added auto start for default docker machine
The docker environment is not bootstrapped after booting a machine and
this often results in a failing rumi run. With this change we try to
automatically start the docker machine and exporting the environment
followed by another attempt to start rumi. Failure in the second case
means that the caller needs to resolve it on her own.